Revert "Looking into the rising temperature problem"
This reverts commit f7969b2c871e1c992ed12e6c25a598e97134dcd7.

+++ lib/rocket/Sensor_MS5607.h
@@ -8,7 +8,6 @@
 
 #include <Wire.h>
 #include "Subsystem.h"
-#include "cstdint"
 
 #define MS5607_RESET 0x1E
 #define MS5607_PROM_READ 0xA0
@@ -57,7 +56,7 @@ private:
     uint32_t D1_pressure{};
     uint32_t D2_temperature{};
 
-    int32_t dT{}, TEMP{}, P{};
+    float dT{}, TEMP{}, P{};
     int64_t OFF{}, SENS{};
 
     uint32_t lastStateChange = 0;
